requirements.txt
# Core async
websockets==12.0
aiohttp==3.9.1
python-dotenv==1.0.0
uvloop==0.19.0
orjson==3.9.10

# Data Processing
numpy==1.26.2
pandas==2.1.4
pyarrow==14.0.1

# Machine Learning
lightgbm==4.1.0
scikit-learn==1.3.2
joblib==1.3.2
optuna==3.4.0

# Blockchain & Polymarket
py-clob-client==0.23.0
web3==6.11.3
eth-account==0.10.0

# Logging
loguru==0.7.2
PyYAML==6.0.1
